requirements.txt
python-dotenv==0.19.1
requests==2.26.0
streamlit==1.0.0